chore: move ProfileModal alongside sibling components
