+++ src/pages/common/Header/Menu/Profile/Profile.tsx
@@ -4,7 +4,7 @@ import type { UserStore } from 'src/stores/User/user.store'
 import { inject, observer } from 'mobx-react'
 import { Flex } from 'theme-ui'
 import { Avatar } from 'src/components/Avatar'
-import { ProfileModal } from 'src/components/ProfileModal/ProfileModal'
+import { ProfileModal } from 'src/pages/common/Header/Menu/ProfileModal/ProfileModal'
 import MenuMobileLink from 'src/pages/common/Header/Menu/MenuMobile/MenuMobileLink'
 import ProfileButtons from './ProfileButtons'
 import { MenuMobileLinkContainer } from '../MenuMobile/MenuMobilePanel'
